Make version requests usable in async contexts
Three constituent parts of the synchronous
`RequesterContext::send_receive_spdm_version` method were split into their own
methods:

 1. Encoding of a message into a buffer,
 2. Tracking the sent message in `runtime_info`
 3. Handling a response message

Creating the send buffer, sending the message, and receiving the message are
still done inside the send_receive_spdm_version method, since that is a
synchronous wrapper method.

Code using async IO or that performs blocking IO in a different, non-immediate,
context should instead call the three constituent methods rather than the
wrapper.

There are a few open design questions here:

* It appears the tracking of sent messages is always done after a message send.
This is probably just to indicate "in-flight" messages for debugging and
fuzzing. It would be nice if we could simplify the async io case by moving the
tracking to before the send. This would just change the semantics to mean that
the last message in the buffer is about to be sent or was sent. Considering the
messages can fail to be received, timeout, or error, this seems like a
reasonable tradeoff. However, I'm unclear of the intent here and so the 3rd
method remains and must be called by a user.

* Should the encapsulation code in `RequesterContext::send_message` and
`RequesterContext::receive_message` be broken out into constituent methods, or
should async io calling code just directly use the `SpdmTransportEncap`. The
goal with this change is to minimize the differences between the synchronous and
the asynchronous code paths, and so where any complex logic exists we want to
break it out. It's unclear if the use of `SpdmTransportEncap` is complex enough
to do this, especially considering that for the user level async io case,
there's no need for a fancy encap/decap.